pipeline refactor
